Update RESTful components for the FamilyMember
