+++ src/main/java/com/hu/familyTree/db/repos/FamilyMemberRepos.java
@@ -18,7 +18,6 @@ public interface FamilyMemberRepos extends CrudRepository<FamilyMemberEntity, Lo
 	@SuppressWarnings("unchecked")
 	FamilyMemberEntity save(FamilyMemberEntity familyMemberEntity);
 	
-	
 	@Query("SELECT u FROM FamilyMemberEntity u WHERE u.firstName = ?1 AND u.lastName = ?2")
 	List<FamilyMemberEntity> findByFirstNameLastName(String firstname, String lastname);
